Conflicts:
	Projects/ClientApp/Src/ClientApplication.cpp
	Projects/Gameplay/Src/EntityFactory.cpp
	Projects/Gameplay/Src/LoadMeshSystem.cpp
	Projects/Gameplay/Src/PacketType.h
	Projects/Graphics/Src/GraphicsWrapper.h
	Projects/ServerApp/Src/ServerApplication.cpp

+++ Projects/Gameplay/Src/ClientPickingSystem.cpp
@@ -80,7 +80,7 @@ void ClientPickingSystem::processEntities(const vector<Entity*>& p_entities)
 
 		double x = input->getCursor()->getX();
 		double y = -input->getCursor()->getY();
-		AglVector4 targetNDC( x, y, 1.0f, 1.0f );
+		AglVector4 targetNDC( (float)x, (float)y, 1.0f, 1.0f );
 		targetNDC.transform(inv);
 		targetNDC /= targetNDC.w;
 		AglVector3 target(targetNDC.x, targetNDC.y, targetNDC.z);
